Ship logs from dedicated Syslog CloudWatch log group (#13)
We don't want the Vector server logs combined with the actual syslogs
being sent. Create a dedicated log group for the syslog messages.

Create a new Functionbeat shipper to process this log stream and update
tags so syslogs can be identified.

Vector server needs to be updated to reference this new log group.