Increase threshold of 500 alarms on API Gateway
The integration between API Gateway and SQS is owned by AWS, seeing
intermittent (once or twice a day) where API Gateway fails a usage plan.
This request is retried and we don't want alarms to go off when there
has been a single 500 integration error. Bump this threshold to 10 to
indicate serious failure.