Commented out VPC peering
error in destroy plan
"Error: Invalid count argument

  on modules/vpc_peering/main.tf line 12, in resource "aws_route" "peering_route":
  12:   count = var.enabled ? length(var.source_route_table_ids) : 0

The "count" value depends on resource attributes that cannot be determined
until apply, so Terraform cannot predict how many instances will be created.
To work around this, use the -target argument to first apply only the
resources that the count depends on.
"